extracting level tutorial
