+++ Assets/Scripts/Objects/Minion.cs
@@ -150,8 +150,8 @@ public class Minion : MonoBehaviour {
 		if (!this.clicksEnabled)
 			return;
 
-		if (LevelManager.singleton.showingTutorials)
-			LevelManager.singleton.MinionClickedInTutorial (this);
+		if (LevelManager.singleton.tutorial != null)
+			LevelManager.singleton.tutorial.MinionClickedInTutorial (this);
 
 		this.hero.PickUpMinion (this);
 	}
