+++ Assets/Scripts/SynthesisTutorial.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 736f0b53c6870844bb9ab287009bdbd1
+timeCreated: 1518742163
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
